+++ core/commons/docker/docker_manager.go
@@ -14,63 +14,131 @@ import (
 	"github.com/sirupsen/logrus"
 	"io"
 	"io/ioutil"
-	"strconv"
 	"time"
 )
 
+/*
+WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING
+
+This manager is used on a per-test basis. Because tests can run in parallel but we need to pretty-print
+each test's logs in a single block, we need to have a seprate logger per test. As such, this class takes in a
+logrus.Logger, and *all log messages should be sent through this logger rather than the systemwide logger!!!*
+
+No logrus.Info, logrus.Debug, etc. calls should happen in this file - only manager.log.Info, manager.log.Debug, etc.!
+
+WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING
+ */
+
 const (
-	DOCKER_NETWORK_NAME ="kurtosis-bridge"
+	DOCKER_NETWORK_DRIVER = "bridge"
 )
 
 type DockerManager struct {
+	// WARNING: This log should be used for all log statements - the system-wide logger should NOT be used!
+	log *logrus.Logger
+
+	// TODO Remove this and create a new Context per Docker function call!!! See
+	//  https://golang.org/pkg/context/ and
+	//  https://blog.golang.org/context
+	//  This would be especially useful for doing timeouts!!
 	dockerCtx           context.Context
 	dockerClient        *client.Client
 }
 
-func NewDockerManager(dockerCtx context.Context, dockerClient *client.Client) (dockerManager *DockerManager, err error) {
+func NewDockerManager(log *logrus.Logger, dockerCtx context.Context, dockerClient *client.Client) (dockerManager *DockerManager, err error) {
 	return &DockerManager{
+		log: log,
 		dockerCtx:           dockerCtx,
 		dockerClient:        dockerClient,
 	}, nil
 }
 
-func (manager DockerManager) CreateNetwork(subnetMask string, gatewayIP string) (id string, err error)  {
-	networkId, ok, err := manager.getNetworkId(DOCKER_NETWORK_NAME)
+// TODO Make this function return the networkId - this would save a TON of hassle, because everywhere else in Docker needs
+//  the network ID and we're passing around the name so we have to do a bunch of Docker lookups every time we want the ID
+/*
+Creates a Docker network with the given parameters, doing nothing if a network with that name already exists
+
+Args:
+	name: The name to give the new Docker network
+	subnetMask: The subnet mask of allowed IPs for the Docker network
+	gatewayIP: The IP to give the network gateway
+
+Returns:
+	id: The Docker-managed ID of the network
+ */
+func (manager DockerManager) CreateNetwork(name string, subnetMask string, gatewayIP string) (id string, err error)  {
+	_, found, err := manager.getNetworkId(name)
 	if err != nil {
-		return "", stacktrace.Propagate(err, "Failed to check for network existence.")
+		return "", stacktrace.Propagate(err, "An error occurred checking for existence of network with name %v", name)
 	}
-	// Network already exists - return existing id.
-	if ok {
-		return networkId, nil
+	if found {
+		// We throw an error if the network already exists because we don't know what settings that network was created
+		//  with - likely a completely different subnetMask and gatewayIP
+		return "", stacktrace.NewError("Network with name %v cannot be created because it already exists", name)
 	}
 	ipamConfig := []network.IPAMConfig{{
 		Subnet: subnetMask,
 		Gateway: gatewayIP,
 	}}
-	resp, err := manager.dockerClient.NetworkCreate(manager.dockerCtx, DOCKER_NETWORK_NAME, types.NetworkCreate{
-		Driver: "bridge",
+	resp, err := manager.dockerClient.NetworkCreate(manager.dockerCtx, name, types.NetworkCreate{
+		Driver: DOCKER_NETWORK_DRIVER,
 		IPAM: &network.IPAM{
 			Config: ipamConfig,
 		},
 	})
 	if err != nil {
-		return "", stacktrace.Propagate( err, "Failed to create network %s with subnet %s", DOCKER_NETWORK_NAME, subnetMask)
+		return "", stacktrace.Propagate( err, "Failed to create network %s with subnet %s", name, subnetMask)
 	}
 	return resp.ID, nil
 }
 
-func (manager DockerManager) CreateVolume(volumeName string) (pathOnHost string, err error) {
+// TODO Change this to be removing a network by ID
+/*
+Removes the Docker network with the given name
+
+Args:
+	networkName: Name of Docker network to remove
+ */
+func (manager DockerManager) RemoveNetwork(networkName string) error {
+	networkId, exists, err := manager.getNetworkId(networkName)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting the network ID for network %v", networkName)
+	}
+	if !exists {
+		// No network with that name exists, so nothing to do
+		return nil
+	}
+	if err := manager.dockerClient.NetworkRemove(manager.dockerCtx, networkId); err != nil {
+		return stacktrace.Propagate(err, "An error occurred removing the Docker network with name %v and ID %v", networkName, networkId)
+	}
+	return nil
+}
+
+/*
+Creates a Docker volume identified by the given name.
+
+Args:
+	volumeName: The unique identifier used by Docker to identify this volume (NOTE: at time of writing, Docker doesn't
+		even give volumes IDs - this name is all there is)
+ */
+func (manager DockerManager) CreateVolume(volumeName string) error {
 	volumeConfig := volume.VolumeCreateBody{
 		Name:       volumeName,
 	}
 
-	volume, err := manager.dockerClient.VolumeCreate(manager.dockerCtx, volumeConfig)
+	/*
+	We don't use the return value of VolumeCreate because there's not much useful information on there - Docker doesn't
+	use UUIDs to identify volumes - only the name - so there's no UUID to retrieve, and the volume's Mountpoint (what you'd
+	think would be the path of the volume on the local machine) isn't useful either becuase Docker itself runs inside a VM
+	so *this path is only a path inside the Docker VM* (meaning we can't use it to read/write files). AFAICT, the only way
+	to read/write data to a volume is to mount it in a container. ~ ktoday, 2020-07-01
+	 */
+	_, err := manager.dockerClient.VolumeCreate(manager.dockerCtx, volumeConfig)
 	if err != nil {
-		return "", stacktrace.Propagate(err, "Could not create Docker volume for test controller")
+		return stacktrace.Propagate(err, "Could not create Docker volume for test controller")
 	}
 
-	// TODO this still isn't tested yet; this might not be the right call
-	return volume.Mountpoint, nil
+	return nil
 }
 
 
@@ -86,33 +154,33 @@ Args:
 	bindMounts: mapping of (host file) -> (mountpoint on container) that will be mounted on container startup
  */
 func (manager DockerManager) CreateAndStartContainer(
-	dockerImage string,
-	staticIp string,
-	usedPorts map[int]bool,
-	startCmdArgs []string,
-	envVariables map[string]string,
-	bindMounts map[string]string,
-	volumeMounts map[string]string) (containerIpAddr string, containerId string, err error) {
+			dockerImage string,
+			networkName string,
+			staticIp string,
+			usedPorts map[nat.Port]bool,
+			startCmdArgs []string,
+			envVariables map[string]string,
+			bindMounts map[string]string,
+			volumeMounts map[string]string) (containerIpAddr string, containerId string, err error) {
 
 	imageExistsLocally, err := manager.isImageAvailableLocally(dockerImage)
 	if err != nil {
-		return "", "", stacktrace.Propagate(err, "Failed to check for image availability.")
+		return "", "", stacktrace.Propagate(err, "An error occurred checking for local availability of Docker image %v", dockerImage)
 	}
 
 	if !imageExistsLocally {
 		err = manager.pullImage(dockerImage)
 		if err != nil {
-			return "", "", stacktrace.Propagate(err, "Failed to pull Docker image.")
+			return "", "", stacktrace.Propagate(err, "Failed to pull Docker image %v from remote image repository", dockerImage)
 		}
 	}
 
-	// TODO replace with custom per-test networks
-	_, networkExistsLocally, err := manager.getNetworkId(DOCKER_NETWORK_NAME)
+	_, networkExistsLocally, err := manager.getNetworkId(networkName)
 	if err != nil {
-		return "", "", stacktrace.Propagate(err, "Failed to check for network availability.")
+		return "", "", stacktrace.Propagate(err, "An error occurred checking for the existence of network %v", networkName)
 	}
 	if !networkExistsLocally {
-		return "", "", stacktrace.NewError("Kurtosis Docker network was never created before trying to launch containers. Please call DockerManager.CreateNetwork first.")
+		return "", "", stacktrace.NewError("Kurtosis Docker network %v was never created before trying to launch containers. Please call DockerManager.CreateNetwork first.", networkName)
 	}
 
 	containerConfigPtr, err := manager.getContainerCfg(dockerImage, usedPorts, startCmdArgs, envVariables)
@@ -123,14 +191,13 @@ func (manager DockerManager) CreateAndStartContainer(
 	if err != nil {
 		return "", "", stacktrace.Propagate(err, "Failed to configure host to container mappings from service.")
 	}
-	// TODO probably use a UUID for the network name (and maybe include test name too)
 	resp, err := manager.dockerClient.ContainerCreate(manager.dockerCtx, containerConfigPtr, containerHostConfigPtr, nil, "")
 	if err != nil {
 		return "", "", stacktrace.Propagate(err, "Could not create Docker container from image %v.", dockerImage)
 	}
 	containerId = resp.ID
 
-	err = manager.connectToNetwork(DOCKER_NETWORK_NAME, containerId, staticIp)
+	err = manager.connectToNetwork(networkName, containerId, staticIp)
 	if err != nil {
 		return "","", stacktrace.Propagate(err, "Failed to connect container %s to network.", containerId)
 	}
@@ -151,6 +218,7 @@ func (manager DockerManager) StopContainer(containerId string, timeout *time.Dur
 	return nil
 }
 
+// TODO Take in a Context, which would allow us to time this out easily!!
 /*
 Blocks until the given container exits.
  */
@@ -184,7 +252,7 @@ func (manager DockerManager) isImageAvailableLocally(imageName string) (isAvaila
 	return len(images) > 0, nil
 }
 
-func (manager DockerManager) getNetworkId(networkName string) (networkId string, ok bool, err error) {
+func (manager DockerManager) getNetworkId(networkName string) (networkId string, found bool, err error) {
 	referenceArg := filters.Arg("name", networkName)
 	filters := filters.NewArgs(referenceArg)
 	networks, err := manager.dockerClient.NetworkList(
@@ -220,7 +288,7 @@ func (manager DockerManager) connectToNetwork(networkName string, containerId st
 }
 
 func (manager DockerManager) pullImage(imageName string) (err error) {
-	logrus.Infof("Pulling image %s...", imageName)
+	manager.log.Infof("Pulling image %s...", imageName)
 	out, err := manager.dockerClient.ImagePull(manager.dockerCtx, imageName, types.ImagePullOptions{})
 	if err != nil {
 		return stacktrace.Propagate(err, "Failed to pull image %s", imageName)
@@ -253,7 +321,7 @@ func (manager *DockerManager) getContainerHostConfig(bindMounts map[string]strin
 		bindsList = append(bindsList, volumeName + ":" + containerFilepath)
 	}
 
-	logrus.Debugf("Binds: %v", bindsList)
+	manager.log.Debugf("Binds: %v", bindsList)
 
 	containerHostConfigPtr := &container.HostConfig{
 		Binds: bindsList,
@@ -265,16 +333,12 @@ func (manager *DockerManager) getContainerHostConfig(bindMounts map[string]strin
 // Creates a Docker container representing a service that will listen on ports in the network
 func (manager *DockerManager) getContainerCfg(
 			dockerImage string,
-			usedPorts map[int]bool,
+			usedPorts map[nat.Port]bool,
 			startCmdArgs []string,
 			envVariables map[string]string) (config *container.Config, err error) {
 	portSet := nat.PortSet{}
 	for port, _ := range usedPorts {
-		otherPort, err := nat.NewPort("tcp", strconv.Itoa(port))
-		if err != nil {
-			return nil, stacktrace.Propagate(err, "Could not parse port int.")
-		}
-		portSet[otherPort] = struct{}{}
+		portSet[port] = struct{}{}
 	}
 
 	envVariablesSlice := make([]string, 0, len(envVariables))
@@ -285,7 +349,6 @@ func (manager *DockerManager) getContainerCfg(
 	nodeConfigPtr := &container.Config{
 		Tty: false,
 		Image: dockerImage,
-		// TODO allow modifying of protocol at some point
 		ExposedPorts: portSet,
 		Cmd: startCmdArgs,
 		Env: envVariablesSlice,