+++ core/commons/networks/public_ip_tracker.go
@@ -3,10 +3,12 @@ package networks
 import (
 	"encoding/binary"
 	"github.com/palantir/stacktrace"
+	"github.com/sirupsen/logrus"
 	"net"
 )
 
 type FreeIpAddrTracker struct {
+	log *logrus.Logger
 	subnet *net.IPNet
 	takenIps map[string]bool
 }
@@ -15,7 +17,7 @@ type FreeIpAddrTracker struct {
 Creates a new tracker that will dole out free IP addresses from the given subnet, making sure not to dole out any IPs
 from the list of already-taken IPs
  */
-func NewFreeIpAddrTracker(subnetMask string, alreadyTakenIps []string) (ipAddrTracker *FreeIpAddrTracker, err error) {
+func NewFreeIpAddrTracker(log *logrus.Logger, subnetMask string, alreadyTakenIps []string) (ipAddrTracker *FreeIpAddrTracker, err error) {
 	_, ipv4Net, err := net.ParseCIDR(subnetMask)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Failed to parse subnet %s as CIDR.", subnetMask)
@@ -27,19 +29,33 @@ func NewFreeIpAddrTracker(subnetMask string, alreadyTakenIps []string) (ipAddrTr
 	}
 
 	ipAddrTracker = &FreeIpAddrTracker{
+		log: log,
 		subnet: ipv4Net,
 		takenIps: takenIps,
 	}
 	return ipAddrTracker, nil
 }
 
+// TODO Return IP objects (which are easily convertable to strings) rather than strings themselves
+// TODO rework this entire function to handle IPv6 as well (currently breaks on IPv6)
 func (networkManager FreeIpAddrTracker) GetFreeIpAddr() (ipAddr string, err error){
 	// convert IPNet struct mask and address to uint32
 	// network is BigEndian
 	mask := binary.BigEndian.Uint32(networkManager.subnet.Mask)
 
+	subnetIp := networkManager.subnet.IP
+
+	// The IP can be either 4 bytes or 16 bytes long; we need to handle both!
+	// See https://gist.github.com/ammario/649d4c0da650162efd404af23e25b86b
+	var intIp uint32
+	if len(subnetIp) == 16 {
+		intIp = binary.BigEndian.Uint32(subnetIp[12:16])
+	} else {
+		intIp = binary.BigEndian.Uint32(subnetIp)
+	}
+
 	// We remove the zeroth IP because it's only used for specifying the network itself
-	start := binary.BigEndian.Uint32(networkManager.subnet.IP) + 1
+	start := intIp + 1
 
 	// find the final address
 	finish := (start & mask) | (mask ^ 0xffffffff)