+++ core/commons/networks/service_network_test.go
@@ -1,6 +1,7 @@
 package networks
 
 import (
+	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/kurtosis/commons/services"
 	"os"
 	"testing"
@@ -12,8 +13,8 @@ type TestService struct {}
 
 // ======================== Test Initializer Core ========================
 type TestInitializerCore struct {}
-func (t TestInitializerCore) GetUsedPorts() map[int]bool {
-	return make(map[int]bool)
+func (t TestInitializerCore) GetUsedPorts() map[nat.Port]bool {
+	return make(map[nat.Port]bool)
 }
 
 func (t TestInitializerCore) GetServiceFromIp(ipAddr string) services.Service {
@@ -56,7 +57,7 @@ func getTestCheckerCore() services.ServiceAvailabilityCheckerCore {
 
 // ======================== Tests ========================
 func TestDisallowingNonexistentConfigs(t *testing.T) {
-	builder := NewServiceNetworkBuilder("test", nil, nil, "test", "/foo/bar")
+	builder := NewServiceNetworkBuilder("test", nil, "test-network", nil, "test", "/foo/bar")
 	network := builder.Build()
 	_, err := network.AddService(0, 0, make(map[int]bool))
 	if err == nil {
@@ -66,7 +67,7 @@ func TestDisallowingNonexistentConfigs(t *testing.T) {
 
 func TestDisallowingNonexistentDependencies(t *testing.T) {
 	configId := 0
-	builder := NewServiceNetworkBuilder("test", nil, nil, "test", "/foo/bar")
+	builder := NewServiceNetworkBuilder("test", nil, "test-network", nil, "test", "/foo/bar")
 	err := builder.AddTestImageConfiguration(configId, getTestInitializerCore(), getTestCheckerCore())
 	if err != nil {
 		t.Fatal("Adding a configuration shouldn't fail")