+++ core/commons/services/service_initializer_core.go
@@ -1,13 +1,16 @@
 package services
 
-import "os"
+import (
+	"github.com/docker/go-connections/nat"
+	"os"
+)
 
 // TODO When Go has generics, parameterize this to be <N, S extends N> where S is the
 //  specific service interface and N represents the interface that every node on the network has
 // Contains configuration determining what type of objects the ServiceInitializer will produce
 // This is implicitly a DockerContainerServiceFactoryConfig; we could abstract it easily if we wanted other foundations for services
 type ServiceInitializerCore interface {
-	GetUsedPorts() map[int]bool
+	GetUsedPorts() map[nat.Port]bool
 
 	// TODO when Go gets generics, make the type of 'dependencies' to be []N
 	// If Go had generics, dependencies should be of type []T